+++ config/util.go
@@ -28,7 +28,7 @@ func LoadConfig(path string) (cfg *Config, err error) {
 		cfg.RPCUser = v.GetString("rpc-user")
 		cfg.RPCPass = v.GetString("rpc-pass")
 		cfg.ShamireHost = v.GetString("shamir-host")
-		cfg.Confirmations = v.GetInt64("confirmations")
+		cfg.Confirmations = v.GetInt("confirmations")
 		cfg.FundAddress = v.GetString("fund-address")
 		cfg.Asset = v.GetString("asset")
 		return
